Refactor sort program factory functions into separate files
Split sort_program_factory.cpp into separate files for better code organization:
- sort_single_row_single_core_program_factory.cpp/.hpp
- sort_cross_core_data_exchange_program_factory.cpp/.hpp
- sort_single_row_multi_core_program_factory.cpp/.hpp

Updated CMakeLists.txt to include new source files and simplified main header to include specialized headers.
