+++ ttnn/cpp/ttnn/operations/data_movement/sort/device/sort_cross_core_data_exchange_program_factory.cpp
@@ -1,0 +1,426 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "sort_cross_core_data_exchange_program_factory.hpp"
+
+#include <tt-metalium/host_api.hpp>
+#include <tt-metalium/constants.hpp>
+#include <tt-metalium/util.hpp>
+
+#include <cmath>
+#include <cstdint>
+
+namespace ttnn::operations::data_movement::sort::program {
+
+// SortProgramFactoryCrossCoreDataExchange - single row, multi core with processing multiple tiles on one core with
+// cross core data exchange
+SortProgramFactoryCrossCoreDataExchange::cached_program_t SortProgramFactoryCrossCoreDataExchange::create(
+    const operation_attributes_t& attributes, const tensor_args_t& tensor_args, tensor_return_value_t& output_tensors) {
+    // Program config
+    tt::tt_metal::Program program{};
+
+    const tt::DataFormat input_tensor_cb_data_format =
+        tt::tt_metal::datatype_to_dataformat_converter(tensor_args.input_tensor.dtype());
+    const tt::DataFormat value_tensor_cb_data_format =
+        tt::tt_metal::datatype_to_dataformat_converter(output_tensors.at(0).dtype());
+    const tt::DataFormat index_tensor_cb_data_format =
+        tt::tt_metal::datatype_to_dataformat_converter(output_tensors.at(1).dtype());
+    const tt::DataFormat packer_unpacker_sync_cb_data_format = tt::DataFormat::Float16_b;
+
+    const uint32_t input_tensor_tile_size = tile_size(input_tensor_cb_data_format);
+    const uint32_t value_tensor_tile_size = tile_size(value_tensor_cb_data_format);
+    const uint32_t index_tensor_tile_size = tile_size(index_tensor_cb_data_format);
+    const uint32_t packer_unpacker_sync_tile_size = tile_size(packer_unpacker_sync_cb_data_format);
+
+    auto input_buffer = tensor_args.input_tensor.buffer();
+    auto value_buffer = output_tensors.at(0).buffer();
+    auto index_buffer = output_tensors.at(1).buffer();
+
+    const bool input_tensor_is_dram = input_buffer->buffer_type() == tt::tt_metal::BufferType::DRAM;
+    const bool value_tensor_is_dram = value_buffer->buffer_type() == tt::tt_metal::BufferType::DRAM;
+    const bool index_tensor_is_dram = index_buffer->buffer_type() == tt::tt_metal::BufferType::DRAM;
+
+    const auto tile_width = tensor_args.input_tensor.tensor_spec().tile().get_width();
+    const auto tile_height = tensor_args.input_tensor.tensor_spec().tile().get_height();
+    const auto tile_hw = tile_width * tile_height;
+
+    const uint32_t num_input_tiles = tensor_args.input_tensor.physical_volume() / tile_hw;
+    const uint32_t num_value_tiles = output_tensors.at(0).physical_volume() / tile_hw;
+
+    const auto input_shape = tensor_args.input_tensor.padded_shape();
+    const uint32_t Ht = (input_shape[0] * input_shape[1] * input_shape[2]) / tile_height;
+    const uint32_t Wt = input_shape[3] / tile_width;
+
+    // Calculate the number of cores available for computation
+    const auto device = tensor_args.input_tensor.device();
+    const auto compute_with_storage_grid_size = device->compute_with_storage_grid_size();
+    const uint32_t total_number_of_cores_physical = compute_with_storage_grid_size.y * compute_with_storage_grid_size.x;
+    const uint32_t total_number_of_cores_virtual = rounddown_pow2(total_number_of_cores_physical);
+    uint32_t number_of_tiles_per_core = get_number_of_tiles_per_core(
+        total_number_of_cores_virtual,
+        Wt,
+        tensor_args.input_tensor.dtype(),
+        output_tensors.at(1).dtype(),
+        CrossCoreDataExchangeSortSlicingStrategy::USE_AS_MANY_CORES);
+    number_of_tiles_per_core = std::min(number_of_tiles_per_core, Wt);
+
+    // Calculate the number of cores utilized based on the input tensor shape
+    const uint32_t all_core_utilization_count = (Wt + number_of_tiles_per_core - 1) / number_of_tiles_per_core;
+
+    TT_FATAL(
+        all_core_utilization_count <= total_number_of_cores_virtual,
+        "All core utilization count exceeds total number of cores. Utilized cores: {}, Total cores: {}",
+        all_core_utilization_count,
+        total_number_of_cores_virtual);
+
+    // uint32 index tensor support
+    const bool is_32_bit_data = index_tensor_cb_data_format == tt::DataFormat::UInt32;
+
+    /**
+     * Calculates the core range based on the number of work units (all_core_utilization_count) and the total number of
+     * available cores in the device's compute grid. The core range determines which cores will be utilized for
+     * computation.
+     *
+     * The calculation works as follows:
+     * 1. If all available cores are needed (all_core_utilization_count == total_number_of_cores), the core range covers
+     * the entire grid.
+     * 2. Otherwise, the number of rows (core_grid_calculated_rows_number) and columns
+     * (core_grid_calculated_columns_number) required to cover all_core_utilization_count are calculated based on the
+     * grid dimensions.
+     *    - If both rows and columns are zero, only a single core is used.
+     *    - If only rows are zero, the core range is set to cover the required number of columns in the first row.
+     *    - Otherwise, the core range is set to cover the required rows, and if there are remaining columns,
+     *      an additional range is added to cover those columns in the next row.
+     *
+     * The resulting core range is represented as a `CoreRangeSet`, which may consist of one or more `CoreRange`
+     * objects depending on the configuration.
+     */
+    CoreRangeSet core_range;
+    if (all_core_utilization_count == total_number_of_cores_physical) {
+        // All cores used
+        core_range = CoreRangeSet(
+            CoreRange({0, 0}, {compute_with_storage_grid_size.x - 1, compute_with_storage_grid_size.y - 1}));
+    } else if (all_core_utilization_count == total_number_of_cores_virtual) {
+        const uint32_t core_grid_calculated_rows_number =
+            all_core_utilization_count / compute_with_storage_grid_size.x - 1;
+        const uint32_t core_grid_calculated_columns_number =
+            all_core_utilization_count % compute_with_storage_grid_size.x;
+        // All virtual cores used
+        core_range =
+            CoreRangeSet(CoreRange({0, 0}, {compute_with_storage_grid_size.x - 1, core_grid_calculated_rows_number}));
+        if (core_grid_calculated_columns_number != 0) {
+            const CoreRange additional_range(
+                {0, core_grid_calculated_rows_number + 1},
+                {core_grid_calculated_columns_number - 1, core_grid_calculated_rows_number + 1});
+            core_range = core_range.merge(CoreRangeSet(additional_range));
+        }
+    } else {
+        const uint32_t core_grid_calculated_rows_number = all_core_utilization_count / compute_with_storage_grid_size.x;
+        const uint32_t core_grid_calculated_columns_number =
+            all_core_utilization_count % compute_with_storage_grid_size.x;
+
+        if (core_grid_calculated_rows_number == 0 && core_grid_calculated_columns_number == 0) {
+            // Only one core used
+            core_range = CoreRangeSet(CoreCoord({0, 0}));
+        } else if (core_grid_calculated_rows_number == 0) {
+            // Only cores from first row used
+            core_range = CoreRangeSet(CoreRange({0, 0}, {core_grid_calculated_columns_number - 1, 0}));
+        } else {
+            // Rows and columns used
+            core_range = CoreRangeSet(
+                CoreRange({0, 0}, {compute_with_storage_grid_size.x - 1, core_grid_calculated_rows_number - 1}));
+            if (core_grid_calculated_columns_number != 0) {
+                const CoreRange additional_range(
+                    {0, core_grid_calculated_rows_number},
+                    {core_grid_calculated_columns_number - 1, core_grid_calculated_rows_number});
+                core_range = core_range.merge(CoreRangeSet(additional_range));
+            }
+        }
+    }
+
+    // Lookup tensor data with physical core coordinates
+    std::vector<uint32_t> physical_core_lookup_table_data;
+    for (const auto& core_range : core_range.ranges()) {
+        for (const auto& core_coord : core_range) {
+            const auto physical_core = device->worker_core_from_logical_core(core_coord);
+            physical_core_lookup_table_data.emplace_back(physical_core.x);
+            physical_core_lookup_table_data.emplace_back(physical_core.y);
+        }
+    }
+    const TensorSpec physical_core_lookup_table_spec(
+        ttnn::Shape{1, physical_core_lookup_table_data.size()},
+        TensorLayout{DataType::UINT32, PageConfig{Layout::ROW_MAJOR}, MemoryConfig()});
+    Tensor physical_core_lookup_table_tensor =
+        Tensor::from_vector(std::move(physical_core_lookup_table_data), physical_core_lookup_table_spec);
+    physical_core_lookup_table_tensor = physical_core_lookup_table_tensor.to_device(device);
+    const auto physical_core_lookup_table_tensor_buffer = physical_core_lookup_table_tensor.buffer();
+    const tt::DataFormat physical_core_lookup_table_cb_data_format =
+        tt::tt_metal::datatype_to_dataformat_converter(physical_core_lookup_table_tensor.dtype());
+    const uint32_t physical_core_lookup_table_tile_size = tile_size(physical_core_lookup_table_cb_data_format);
+    const bool physical_core_lookup_table_is_dram =
+        physical_core_lookup_table_tensor_buffer->buffer_type() == tt::tt_metal::BufferType::DRAM;
+
+    // Circular buffers
+    constexpr uint32_t cb_scale_factor = 2;
+
+    constexpr uint32_t input_tensor_cb_index = tt::CBIndex::c_0;
+    const tt::tt_metal::CircularBufferConfig input_tensor_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * input_tensor_tile_size, {{input_tensor_cb_index, input_tensor_cb_data_format}})
+            .set_page_size(input_tensor_cb_index, input_tensor_tile_size);
+    const auto cb_input_tensor = tt::tt_metal::CreateCircularBuffer(program, core_range, input_tensor_cb_config);
+
+    constexpr uint32_t index_tensor_cb_index = tt::CBIndex::c_1;
+    const tt::tt_metal::CircularBufferConfig index_tensor_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * index_tensor_tile_size, {{index_tensor_cb_index, index_tensor_cb_data_format}})
+            .set_page_size(index_tensor_cb_index, index_tensor_tile_size);
+    const auto cb_index_tensor = tt::tt_metal::CreateCircularBuffer(program, core_range, index_tensor_cb_config);
+
+    constexpr uint32_t input_tensor_transposed_cb_index = tt::CBIndex::c_2;
+    const tt::tt_metal::CircularBufferConfig input_tensor_transposed_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            number_of_tiles_per_core * input_tensor_tile_size,
+            {{input_tensor_transposed_cb_index, input_tensor_cb_data_format}})
+            .set_page_size(input_tensor_transposed_cb_index, input_tensor_tile_size);
+    const auto cb_input_tensor_transposed =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, input_tensor_transposed_cb_config);
+
+    constexpr uint32_t index_tensor_transposed_cb_index = tt::CBIndex::c_3;
+    const tt::tt_metal::CircularBufferConfig index_tensor_transposed_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            number_of_tiles_per_core * index_tensor_tile_size,
+            {{index_tensor_transposed_cb_index, index_tensor_cb_data_format}})
+            .set_page_size(index_tensor_transposed_cb_index, index_tensor_tile_size);
+    const auto cb_index_tensor_transposed =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, index_tensor_transposed_cb_config);
+
+    constexpr uint32_t value_tensor_cb_index = tt::CBIndex::c_4;
+    const tt::tt_metal::CircularBufferConfig value_tensor_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * value_tensor_tile_size, {{value_tensor_cb_index, value_tensor_cb_data_format}})
+            .set_page_size(value_tensor_cb_index, index_tensor_tile_size);
+    const auto cb_value_tensor = tt::tt_metal::CreateCircularBuffer(program, core_range, value_tensor_cb_config);
+
+    constexpr uint32_t index_tensor_output_cb_index = tt::CBIndex::c_5;
+    const tt::tt_metal::CircularBufferConfig index_tensor_output_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * index_tensor_tile_size, {{index_tensor_output_cb_index, index_tensor_cb_data_format}})
+            .set_page_size(index_tensor_output_cb_index, index_tensor_tile_size);
+    const auto cb_index_tensor_output =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, index_tensor_output_cb_config);
+
+    constexpr uint32_t value_tensor_intermediate_cb_index = tt::CBIndex::c_6;
+    const tt::tt_metal::CircularBufferConfig value_tensor_intermediate_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * value_tensor_tile_size,
+            {{value_tensor_intermediate_cb_index, value_tensor_cb_data_format}})
+            .set_page_size(value_tensor_intermediate_cb_index, index_tensor_tile_size);
+    const auto cb_value_intermediate_tensor =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, value_tensor_intermediate_cb_config);
+
+    constexpr uint32_t index_tensor_intermediate_cb_index = tt::CBIndex::c_7;
+    const tt::tt_metal::CircularBufferConfig index_tensor_intermediate_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * index_tensor_tile_size,
+            {{index_tensor_intermediate_cb_index, index_tensor_cb_data_format}})
+            .set_page_size(index_tensor_intermediate_cb_index, index_tensor_tile_size);
+    const auto cb_index_tensor_intermediate =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, index_tensor_intermediate_cb_config);
+
+    constexpr uint32_t value_tensor_peer_cb_index = tt::CBIndex::c_8;
+    const tt::tt_metal::CircularBufferConfig value_tensor_peer_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * value_tensor_tile_size, {{value_tensor_peer_cb_index, value_tensor_cb_data_format}})
+            .set_page_size(value_tensor_peer_cb_index, index_tensor_tile_size);
+    const auto cb_value_peer_tensor =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, value_tensor_peer_cb_config);
+
+    constexpr uint32_t index_tensor_peer_cb_index = tt::CBIndex::c_9;
+    const tt::tt_metal::CircularBufferConfig index_tensor_peer_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            cb_scale_factor * index_tensor_tile_size, {{index_tensor_peer_cb_index, index_tensor_cb_data_format}})
+            .set_page_size(index_tensor_peer_cb_index, index_tensor_tile_size);
+    const auto cb_index_tensor_peer =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, index_tensor_peer_cb_config);
+
+    constexpr uint32_t physical_core_lookup_table_cb_index = tt::CBIndex::c_10;
+    const tt::tt_metal::CircularBufferConfig physical_core_lookup_table_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            physical_core_lookup_table_tile_size,
+            {{physical_core_lookup_table_cb_index, physical_core_lookup_table_cb_data_format}})
+            .set_page_size(physical_core_lookup_table_cb_index, physical_core_lookup_table_tile_size);
+    const auto cb_physical_core_lookup_table =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, physical_core_lookup_table_cb_config);
+
+    constexpr uint32_t packer_unpacker_sync_cb_index = tt::CBIndex::c_11;
+    const tt::tt_metal::CircularBufferConfig packer_unpacker_sync_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            packer_unpacker_sync_tile_size, {{packer_unpacker_sync_cb_index, packer_unpacker_sync_cb_data_format}})
+            .set_page_size(packer_unpacker_sync_cb_index, packer_unpacker_sync_tile_size);
+    const auto cb_packer_unpacker_sync =
+        tt::tt_metal::CreateCircularBuffer(program, core_range, packer_unpacker_sync_cb_config);
+
+    // Semaphores
+    const uint32_t semaphore_exchange_readers = CreateSemaphore(program, core_range, 0);
+    const uint32_t semaphore_exchange_writers = CreateSemaphore(program, core_range, 0);
+    const uint32_t semaphore_barrier = CreateSemaphore(program, core_range, 0);
+
+    // Kernels
+    const std::vector<uint32_t> reader_compile_time_args = {
+        compute_with_storage_grid_size.x,
+        compute_with_storage_grid_size.y,
+        input_tensor_cb_index,
+        index_tensor_output_cb_index,
+        value_tensor_intermediate_cb_index,
+        index_tensor_intermediate_cb_index,
+        value_tensor_peer_cb_index,
+        index_tensor_peer_cb_index,
+        physical_core_lookup_table_cb_index,
+        input_tensor_is_dram,
+        index_tensor_is_dram,
+        physical_core_lookup_table_is_dram,
+        Ht,
+        Wt,
+        number_of_tiles_per_core,
+        all_core_utilization_count,
+        !attributes.descending,
+        semaphore_exchange_readers,
+        semaphore_barrier,
+    };
+    const std::string reader_kernel_path =
+        "ttnn/cpp/ttnn/operations/data_movement/sort/device/kernels/dataflow/reader_cross_core_data_exchange.cpp";
+    tt::tt_metal::KernelHandle reader_kernel_id = tt::tt_metal::CreateKernel(
+        program, reader_kernel_path, core_range, tt::tt_metal::ReaderDataMovementConfig{reader_compile_time_args});
+    SetRuntimeArgs(
+        program,
+        reader_kernel_id,
+        core_range,
+        {input_buffer->address(), index_buffer->address(), physical_core_lookup_table_tensor_buffer->address()});
+
+    const std::vector<uint32_t> writer_compile_time_args = {
+        compute_with_storage_grid_size.x,
+        compute_with_storage_grid_size.y,
+        index_tensor_cb_index,
+        value_tensor_cb_index,
+        value_tensor_peer_cb_index,
+        physical_core_lookup_table_cb_index,
+        value_tensor_is_dram,
+        Wt,
+        Ht,
+        number_of_tiles_per_core,
+        total_number_of_cores_virtual,
+        semaphore_exchange_readers,
+        static_cast<uint32_t>(is_32_bit_data)};
+    const std::string writer_kernel_path =
+        "ttnn/cpp/ttnn/operations/data_movement/sort/device/kernels/dataflow/writer_cross_core_data_exchange.cpp";
+    tt::tt_metal::KernelHandle writer_kernel_id = tt::tt_metal::CreateKernel(
+        program, writer_kernel_path, core_range, tt::tt_metal::WriterDataMovementConfig{writer_compile_time_args});
+    SetRuntimeArgs(program, writer_kernel_id, core_range, {value_buffer->address()});
+
+    const std::vector<uint32_t> compute_compile_time_args = {
+        compute_with_storage_grid_size.x,
+        compute_with_storage_grid_size.y,
+        Ht,
+        Wt,
+        number_of_tiles_per_core,
+        all_core_utilization_count,
+        !attributes.descending,
+        input_tensor_cb_index,
+        index_tensor_cb_index,
+        input_tensor_transposed_cb_index,
+        index_tensor_transposed_cb_index,
+        value_tensor_cb_index,
+        index_tensor_output_cb_index,
+        value_tensor_intermediate_cb_index,
+        index_tensor_intermediate_cb_index,
+        value_tensor_peer_cb_index,
+        index_tensor_peer_cb_index,
+        packer_unpacker_sync_cb_index,
+    };
+    const std::string compute_kernel_path =
+        "ttnn/cpp/ttnn/operations/data_movement/sort/device/kernels/compute/sort_cross_core_data_exchange.cpp";
+    tt::tt_metal::KernelHandle compute_kernel_id = tt::tt_metal::CreateKernel(
+        program,
+        compute_kernel_path,
+        core_range,
+        tt::tt_metal::ComputeConfig{.fp32_dest_acc_en = is_32_bit_data, .compile_args = compute_compile_time_args});
+
+    return {std::move(program), {reader_kernel_id, compute_kernel_id, writer_kernel_id, core_range}};
+}
+
+void SortProgramFactoryCrossCoreDataExchange::override_runtime_arguments(
+    cached_program_t& cached_program,
+    const operation_attributes_t& attributes,
+    const tensor_args_t& tensor_args,
+    tensor_return_value_t& output_tensors) {
+    const auto input_tensor_buffer = tensor_args.input_tensor.buffer();
+    const auto value_tensor_buffer = output_tensors.at(0).buffer();
+    const auto index_tensor_buffer = output_tensors.at(1).buffer();
+    const auto device = tensor_args.input_tensor.device();
+
+    // Lookup tensor data with physical core coordinates
+    std::vector<uint32_t> physical_core_lookup_table_data;
+    for (const auto& core_range : cached_program.shared_variables.core_range_set.ranges()) {
+        for (const auto& core_coord : core_range) {
+            const auto physical_core = device->worker_core_from_logical_core(core_coord);
+            physical_core_lookup_table_data.emplace_back(physical_core.x);
+            physical_core_lookup_table_data.emplace_back(physical_core.y);
+        }
+    }
+    const TensorSpec physical_core_lookup_table_spec(
+        ttnn::Shape{1, physical_core_lookup_table_data.size()},
+        TensorLayout{DataType::UINT32, PageConfig{Layout::ROW_MAJOR}, MemoryConfig()});
+    Tensor physical_core_lookup_table_tensor =
+        Tensor::from_vector(std::move(physical_core_lookup_table_data), physical_core_lookup_table_spec);
+    physical_core_lookup_table_tensor = physical_core_lookup_table_tensor.to_device(device);
+
+    // Update runtime args
+    for (const auto& core_range : cached_program.shared_variables.core_range_set.ranges()) {
+        for (const auto& core_coord : core_range) {
+            auto& reader_runtime_args =
+                GetRuntimeArgs(cached_program.program, cached_program.shared_variables.reader_kernel_id, core_coord);
+            reader_runtime_args[0] = input_tensor_buffer->address();
+            reader_runtime_args[1] = index_tensor_buffer->address();
+            reader_runtime_args[2] = physical_core_lookup_table_tensor.buffer()->address();
+
+            auto& writer_runtime_args =
+                GetRuntimeArgs(cached_program.program, cached_program.shared_variables.writer_kernel_id, core_coord);
+            writer_runtime_args[0] = value_tensor_buffer->address();
+        }  // core_coord loop
+    }  // core_range loop
+}
+
+uint32_t SortProgramFactoryCrossCoreDataExchange::get_number_of_tiles_per_core(
+    uint32_t total_number_of_cores,
+    uint32_t Wt,
+    const DataType& input_dtype,
+    const DataType& index_dtype,
+    CrossCoreDataExchangeSortSlicingStrategy slicing_strategy) {
+    switch (slicing_strategy) {
+        case CrossCoreDataExchangeSortSlicingStrategy::USE_AS_MANY_CORES: {
+            constexpr uint32_t MIN_TILES_PER_CORE = 2;
+            return std::max(Wt / total_number_of_cores, MIN_TILES_PER_CORE);
+        }
+        case CrossCoreDataExchangeSortSlicingStrategy::FILL_CORES_FIRST:
+        default: {
+            if (input_dtype == DataType::FLOAT32 || input_dtype == DataType::UINT32 || input_dtype == DataType::INT32 ||
+                index_dtype == DataType::INT32 || index_dtype == DataType::UINT32) {
+                return 64;
+            }
+            break;
+        }
+    }
+
+    return 128;
+}
+
+uint32_t SortProgramFactoryCrossCoreDataExchange::rounddown_pow2(uint32_t n) {
+    if (n == 0) {
+        return 0;
+    }
+    return 1 << (31 - std::countl_zero(n));
+}
+
+}  // namespace ttnn::operations::data_movement::sort::program
